requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
slowapi>=0.1.9
python-dotenv>=1.0.0
structlog>=24.1.0
pydantic>=2.7.0
httpx>=0.27.0
